+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Retry incoming activities only on fetcher errors.
 - Don't fetch object of FEP-1b12 announced `Create` and `Update` activities.
 - Update actor reachability statuses with a single query.
+- Don't publish post if `Idempotency-Key` header value is reused.
 
 ### Fixed